Procedural map
We've changed from having just a small wall of 3 tiles, to now creating an entire map randomly generated with room and tunnels connecting them and then the player is placed in the center of one of those rooms. Walls are dark and are blocked from passage, light color is "ground" which creates paths that can be traveled.