FOV rendering
Finished the Field of View Algorithm.  Created 2 light colors for when the player can see directly into their line of sight.  Then created a radius of 'light' of 10 spaces.  Now map is 'hidden' unless in FOV as light colors or 'explored' as dark colors.  Also set Entity generation to include whether in FOV or not when drawn.